+++ src/nn/model/layers/node.h
@@ -9,7 +9,7 @@
 #include "utils/mem/mem-management.h"
 
 typedef struct nn_Node {
-  unsigned int num_weights;
+  size_t num_weights;
   // weights
   double* weights;
   double* d_weights;
@@ -25,7 +25,7 @@ typedef struct nn_Node {
   double d_value;
 } nn_Node;
 
-nn_Node* _nn_createNode(unsigned int num_weights);
+nn_Node* _nn_createNode(size_t num_weights);
 void _nn_freeNode(nn_Node* node);
 
 #endif
